feat: video feed architecture consolidation and cleanup
- Remove unused video_feed_provider_v2.dart (old ChangeNotifier pattern)
- Remove unused video_player_widget.dart (replaced by video_playback_widget)
- Remove unused video_manager_provider.dart (migration wrapper)
- Keep modern Riverpod implementations and unified video architecture
- Delete associated test files for removed components

Part of codebase cleanup plan Week 2 tasks

+++ mobile/lib/scripts/bulk_thumbnail_generator.dart
@@ -221,7 +221,7 @@ Examples:
     Map<String, dynamic> options
   ) async {
     final isDryRun = options['dryRun'] == true;
-    final batchSizeToUse = options['batchSize'] ?? batchSize;
+    final batchSizeToUse = (options['batchSize'] ?? batchSize) as int;
     final timeOffset = options['timeOffset'] ?? 2.5;
     
     if (isDryRun) {
